Explicitly request a qemu pc-1.0 machine type.
As of libvirt 0.9.8-2ubuntu17.2, it now automatically "upgrades" you
to latest. But that latest version chosen does not seem to be
coordinated with what qemu actually supports, and qemu ends up
erroring out. Explicit specify a version to avoid this logic, and pick
a version that actually works with the current version of qemu
(1.0+noroms-0ubuntu14.1).