Support human-friendly input for size units.
This supports both SI prefixes (MB etc) and power-of-two engineering
conventions (MiB etc). Additionally, just the prefix (M etc) is
accepted as shorthand for power-of-two units (that is, M means MiB
etc).

The fields that currently accept these are ``ram`` and ``disk-size``,
for example, in your meta-data file, say:

  downburst:
    ram: 4G
    disk-size: 8G

Warning: the ``ram`` field default unit changes from KiB to bytes!
Adjust your existing meta-data files to include an explicit unit.
This is desirable in order to have all size fields be in bytes, to
avoid the confusion of sequential lines being in different units.

Thanks to Sandon Van Ness <sandon@inktank.com> for the initial code.

Closes: #2818