Added default SSH key support.
Added quite some time ago but forgot to commit. This adds a default
key setting on downburst so not only is the user's ssh key installed
but a 'generic' or 'default' ssh key (which you want on all your
machines) also gets installed. In our sepia lab case this is
teuthology's ssh key and the behavior can be disabled with a command
line arugment of --nokey to downburst.