Added support for creating guests of multiple architecture.
Downburst now supports creating guests that are x86_64/amd64 or i386.
This is done via either a commandline addition to the downburst
command eg. --arch=i386 or in the yaml file with arch: i386. The
default arch is x86_64/amd64. It will take both x86_64 or amd64 as
arguments to use a 64-bit image.