Support to downburst to download Cloudinit Images off ceph.com
Added cloud images to ceph.com (from my own server) and modified
discover.py which used to be able to download ubuntu (only) images
of cloud.ubuntu.com. Unfortunately ubuntu dropped their json
support so the old code was broken to begin with. I now just parsing
an apache directory listing to find out of the image is downloadable
or not.

In the future this should be modified to be more elegent and use
an HTTP parser.

Signed-off-by: Sandon Van Ness <sandon@inktank.com>