Support for creating guests with multiple drives.
Downburst originally only supported creating a guest with a single
boot drive. This change allows the user to specify the amount of
additional drives in the metadata yaml file. This is done with

 additional-disks: $x
 additional-disks-size: '$yG'

where $x is the amount of drives wanted and $y is the size of each
of the the additional drives. If size is not specified it defaults
to 10G. The drives are setup as raw images with fallocate so they
are created fast and should offer the best possible performance
as unlike qcow it will fully allocate the images and they wont be
fragmented or have other negative performance impacts of sparse
images.

Signed-off-by: Sandon Van Ness <sandon@inktank.com>
Reviewed-by: Warren Usui <warren.usui@inktank.com>