Check for corrupted imagefile.
The size thing is fairly arbitrary. It looks like with partial downloaded
files the size will be 0 due to them being root owned. This might change
on libvirt-bin restart on the hosts.

The <50 megabyte thing is more for broken requests which was breaking in
alternate way. Either way this check should be good for removing existing
broken images and re-download.

Signed-off-by: Sandon Van Ness <sandon@inktank.com>